+++ mxm/src/main/java/org/mbari/mxm/client/BuildInfo.java
@@ -1,5 +1,5 @@
 package org.mbari.mxm.client;
 
 public class BuildInfo {
-  public static String getVersion() { return "0.1.1"; }
+  public static String getVersion() { return "0.1.2"; }
 }